requirements.txt
asgiref==3.5.2
Django==4.1.3
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-utils-six==2.0
phonenumbers==8.13.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
martor
gunicorn
